+++ app/api/domain/repository/index.ts
@@ -1,0 +1,1 @@
+export type { UserRepository } from './UserRepository';
