+++ app/api/infrastructure/repositories/index.ts
@@ -1,0 +1,3 @@
+export { SbRsvRepository } from './SbRsvRepository';
+export { SbSpaceRepository } from './SbSpaceRepository';
+export { SupabaseUserRepository } from './SbUserRepository';
